+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Notable `mc-legacy-formatting` changes, tracked in the [keep a changelog](https:
 
 ## [Unreleased]
 
+### Internal
+
+* `SpanIter` no longer keeps track of a `finished` flag as it was unnecessary
+
 ## [0.3.1] - 2020-11-19
 
 ### Changed